[feat] 상세페이지 비슷한 콘텐츠에 샘플 api 데이터 끌어와보면서 1차 redux slice와 getthunk 구조, 받는쪽 selector, dispatch props전달 구조 설계
