+++ src/store/index.js
@@ -1,17 +1,19 @@
 import { configureStore } from '@reduxjs/toolkit';
 // import aboutR from './modules/aboutSlice';
-import contentR from './modules/contentSlice';
+// import contentR from './modules/contentSlice';
 // import memberR from './modules/memberSlice';
 import playerR from './modules/playerSlice';
 // import tmR from './modules/tmSlice';
+import movieR from './modules/movieSlice';
 // import pagenationR from './modules/pagenationSlice';
 
 export const store = configureStore({
     reducer: {
         // aboutR,
-        contentR,
+        // contentR,
         // memberR,
         playerR,
+        movieR,
         // tmR,
         // pagenationR,
     },
