+++ src/store/modules/contentSlice.js
@@ -1,13 +1,13 @@
-import { createSlice } from '@reduxjs/toolkit';
-// import 참조변수 from '../../assets/api/데이터';
+// import { createSlice } from '@reduxjs/toolkit';
+// // import 참조변수 from '../../assets/api/데이터';
 
-const initialState = {};
+// const initialState = {};
 
-export const contentSlice = createSlice({
-  name: 'content',
-  initialState,
-  reducers: {},
-});
+// export const contentSlice = createSlice({
+//   name: 'content',
+//   initialState,
+//   reducers: {},
+// });
 
-export const contentActions = contentSlice.actions;
-export default contentSlice.reducer;
+// export const contentActions = contentSlice.actions;
+// export default contentSlice.reducer;
